requirements.txt
python-telegram-bot>=21.0
anthropic>=0.40.0
supabase>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.35.0
plotly>=5.18.0
pandas>=2.1.0
Pillow>=10.0.0
